worker/requirements.txt
celery
pillow
pytesseract
redis
sqlalchemy
python-dotenv
